fix(cli): invalid plugin template (#8161)
