+++ .changes/fix-plugin-template.md
@@ -1,0 +1,6 @@
+---
+"@tauri-apps/cli": patch:bug
+"tauri-cli": patch:bug
+---
+
+Fix invalid plugin template.
