+++ tooling/cli/templates/plugin/__example-api/tauri-app/package.json
@@ -10,13 +10,14 @@
     "tauri": "tauri"
   },
   "dependencies": {
-    "@tauri-apps/api": "^1.1.0",
+    "@tauri-apps/api": "^2.0.0-alpha.11",
     "tauri-plugin-{{ plugin_name }}-api": "link:../../"
   },
   "devDependencies": {
     "@sveltejs/vite-plugin-svelte": "^1.0.1",
+    "internal-ip": "^7.0.0",
     "svelte": "^3.49.0",
     "vite": "^3.0.2",
-    "@tauri-apps/cli": "^1.1.0"
+    "@tauri-apps/cli": "^2.0.0-alpha.17"
   }
 }
