+++ tooling/cli/templates/plugin/__example-api/tauri-app/src-tauri/Cargo.crate-manifest
@@ -8,6 +8,10 @@ repository = ""
 edition = "2021"
 rust-version = "1.70"
 
+[lib]
+name = "tauri_app_lib"
+crate-type = ["staticlib", "cdylib", "rlib"]
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [build-dependencies]
