+++ tooling/cli/templates/plugin/__example-api/tauri-app/src-tauri/src/lib.rs
@@ -1,0 +1,14 @@
+// Learn more about Tauri commands at https://tauri.app/v1/guides/features/command
+#[tauri::command]
+fn greet(name: &str) -> String {
+    format!("Hello, {}! You've been greeted from Rust!", name)
+}
+
+#[cfg_attr(mobile, tauri::mobile_entry_point)]
+pub fn run() {
+    tauri::Builder::default()
+        .invoke_handler(tauri::generate_handler![greet])
+        .plugin(tauri_plugin_{{ plugin_name_snake_case }}::init())
+        .run(tauri::generate_context!())
+        .expect("error while running tauri application");
+}
