+++ tooling/cli/templates/plugin/__example-api/tauri-app/vite.config.js
@@ -1,5 +1,8 @@
 import { defineConfig } from "vite";
 import { svelte } from "@sveltejs/vite-plugin-svelte";
+import { internalIpV4Sync } from 'internal-ip'
+
+const mobile = !!/android|ios/.exec(process.env.TAURI_ENV_PLATFORM);
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -10,18 +13,13 @@ export default defineConfig({
   clearScreen: false,
   // tauri expects a fixed port, fail if that port is not available
   server: {
+    host: mobile ? "0.0.0.0" : false,
     port: 1420,
     strictPort: true,
+    hmr: mobile ? {
+      protocol: 'ws',
+      host: internalIpV4Sync(),
+      port: 1421
+    } : undefined,
   },
-  // to make use of `TAURI_ENV_DEBUG` and other env variables
-  // https://tauri.studio/v1/api/config#buildconfig.beforedevcommand
-  envPrefix: ["VITE_", "TAURI_ENV_"],
-  build: {
-    // Tauri supports es2021
-    target: ["es2021", "chrome100", "safari13"],
-    // don't minify for debug builds
-    minify: !process.env.TAURI_ENV_DEBUG ? "esbuild" : false,
-    // produce sourcemaps for debug builds
-    sourcemap: !!process.env.TAURI_ENV_DEBUG,
-  },
-});
+})
