+++ tooling/cli/templates/plugin/__example-basic/vanilla/package.json
@@ -9,6 +9,6 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "@tauri-apps/cli": "^1.0.0"
+    "@tauri-apps/cli": "^2.0.0-alpha.17"
   }
 }
