+++ tooling/cli/templates/plugin/__example-basic/vanilla/src-tauri/src/lib.rs
@@ -1,0 +1,7 @@
+#[cfg_attr(mobile, tauri::mobile_entry_point)]
+pub fn run() {
+  tauri::Builder::default()
+    .plugin(tauri_plugin_{{ plugin_name_snake_case }}::init())
+    .run(tauri::generate_context!())
+    .expect("error while running tauri application");
+}
