+++ tooling/cli/templates/plugin/__example-basic/vanilla/src-tauri/src/main.rs
@@ -2,8 +2,5 @@
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
 fn main() {
-  tauri::Builder::default()
-    .plugin(tauri_plugin_{{ plugin_name_snake_case }}::init())
-    .run(tauri::generate_context!())
-    .expect("error while running tauri application");
+  tauri_app_lib::run();
 }
